Renamed basic rules
