+++ src/main/java/edu/rpi/legup/puzzle/skyscrapers/rules/LastVisibleCellBasicRule.java
@@ -9,18 +9,14 @@ import edu.rpi.legup.puzzle.skyscrapers.SkyscrapersBoard;
 import edu.rpi.legup.puzzle.skyscrapers.SkyscrapersCell;
 import edu.rpi.legup.puzzle.skyscrapers.SkyscrapersType;
 
-import java.awt.Point;
 import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.Iterator;
-import java.util.Set;
 
-public class OneEdgeBasicRule extends BasicRule {
+public class LastVisibleCellBasicRule extends BasicRule {
 
-    public OneEdgeBasicRule() {
-        super("SKYS-BASC-0005", "One Edge",
-                "This is the last number for this cell that does not create a visibility contradiction",
-                "edu/rpi/legup/images/skyscrapers/rules/OneEdge.png");
+    public LastVisibleCellBasicRule() {
+        super("SKYS-BASC-0001", "Last Visible Cell",
+                "There is only one cell on this row/col for this number that does not create a visibility contradiction",
+                "edu/rpi/legup/images/skyscrapers/rules/FixedMax.png");
     }
 
     /**
@@ -34,7 +30,7 @@ public class OneEdgeBasicRule extends BasicRule {
      */
     @Override
     public String checkRuleRawAt(TreeTransition transition, PuzzleElement puzzleElement) {
-        //last number for cell based upon preemptive visibility rules
+        //last cell for number based on preemptive visibility rules
         SkyscrapersBoard initialBoard = (SkyscrapersBoard) transition.getParents().get(0).getBoard();
         SkyscrapersCell initCell = (SkyscrapersCell) initialBoard.getPuzzleElement(puzzleElement);
         SkyscrapersBoard finalBoard = (SkyscrapersBoard) transition.getBoard();
@@ -48,19 +44,36 @@ public class OneEdgeBasicRule extends BasicRule {
         boolean viewTemp = initialBoard.getViewFlag();
         initialBoard.setDupeFlag(false);
         initialBoard.setViewFlag(true);
-        NumberForCellCaseRule caseRule = new NumberForCellCaseRule();
-        ArrayList<Board> candidates = caseRule.getCases(initialBoard,puzzleElement);
+        CellForNumberCaseRule caseRule = new CellForNumberCaseRule();
+        ArrayList<Board> XCandidates = caseRule.getCasesFor(initialBoard,initialBoard.getWestClues().get(finalCell.getLocation().y),(Integer)finalCell.getData());
+        ArrayList<Board> YCandidates = caseRule.getCasesFor(initialBoard,initialBoard.getNorthClues().get(finalCell.getLocation().x),(Integer)finalCell.getData());
         initialBoard.setDupeFlag(dupeTemp);
         initialBoard.setViewFlag(viewTemp);
 
-        //check if given value is the only remaining value
-        if(candidates.size() == 1){
-            if(candidates.get(0).getPuzzleElement(puzzleElement).getData() == finalCell.getData()){
-                return null;
-            }
-            return super.getInvalidUseOfRuleMessage() + ": Wrong number in the cell.";
+        System.out.println(XCandidates.size());
+        System.out.println(YCandidates.size());
+
+        //return null if either pass, both messages otherwise
+        String xCheck = candidateCheck(XCandidates,puzzleElement,finalCell);
+        String yCheck = candidateCheck(YCandidates,puzzleElement,finalCell);
+        if(xCheck==null || yCheck==null){
+            return null;
         }
-        return super.getInvalidUseOfRuleMessage() + ":This cell is not forced.";
+        return super.getInvalidUseOfRuleMessage() + "\nRow" + xCheck + "\nCol" + yCheck;
+    }
+
+    //helper to check if candidate list is valid
+    private String candidateCheck(ArrayList<Board> candidates,PuzzleElement puzzleElement, SkyscrapersCell finalCell){
+        if(candidates.size() == 1){
+            if(((SkyscrapersCell) candidates.get(0).getPuzzleElement(puzzleElement)).getType() == SkyscrapersType.Number) {
+                if (candidates.get(0).getPuzzleElement(puzzleElement).getData() == finalCell.getData()) {
+                    return null;
+                }
+                return ": Wrong number in the cell.";
+            }
+            return ": No case for this cell.";
+        }
+        return ": This cell is not forced.";
     }
 
     private boolean isForced(SkyscrapersBoard board, SkyscrapersCell cell) {
@@ -82,22 +95,22 @@ public class OneEdgeBasicRule extends BasicRule {
      */
     @Override
     public Board getDefaultBoard(TreeNode node) {
-        SkyscrapersBoard initialBoard = (SkyscrapersBoard) node.getBoard();
-        SkyscrapersBoard lightUpBoard = (SkyscrapersBoard) node.getBoard().copy();
-        System.out.println(lightUpBoard.getPuzzleElements().size());
-        for (PuzzleElement element : lightUpBoard.getPuzzleElements()) {
+    	SkyscrapersBoard initialBoard = (SkyscrapersBoard) node.getBoard();
+    	SkyscrapersBoard modBoard = (SkyscrapersBoard) node.getBoard().copy();
+    	System.out.println(modBoard.getPuzzleElements().size());
+        for (PuzzleElement element : modBoard.getPuzzleElements()) {
             System.out.println("123");
-            SkyscrapersCell cell = (SkyscrapersCell) element;
+        	SkyscrapersCell cell = (SkyscrapersCell) element;
             if (cell.getType() == SkyscrapersType.UNKNOWN && isForced(initialBoard, cell)) {
                 //cell.setData(SkyscrapersType.BULB.value);
-                lightUpBoard.addModifiedData(cell);
+                modBoard.addModifiedData(cell);
             }
         }
-        if (lightUpBoard.getModifiedData().isEmpty()) {
+        System.out.println(modBoard.getModifiedData().isEmpty());
+        if (modBoard.getModifiedData().isEmpty()) {
             return null;
-        }
-        else {
-            return lightUpBoard;
+        } else {
+            return modBoard;
         }
     }
 }
