+++ src/main/java/edu/rpi/legup/puzzle/skyscrapers/rules/TODO.md
@@ -13,8 +13,9 @@ spreadsheet : https://docs.google.com/spreadsheets/d/1l7aUZtavtysM8dtGnaEIXhBKMR
       - View contains a few of these
     - document utility functions in the reference sheet, COMMENTS!
     - review and identify dead code
-    - replace height/width with size (never not square)
     - remove all these damn print statments (commented ones too if they aren't useful)
+    - Edit to allow blank clues
+    - Display flags somewhere
  5. Flags
     - edit exporter to include flags in xml file format (if needed)
  6. Documentation
